[Silver V] Title: FIFA 월드컵, Time: 48 ms, Memory: 5680 KB -BaekjoonHub

+++ 백준/Silver/4884. FIFA 월드컵/FIFA 월드컵.cs
@@ -13,15 +13,15 @@ namespace hello
                 if (s == "-1 -1 -1 -1") break;
                 string[] ss = s.Split();
                 long[] num = Array.ConvertAll(ss, long.Parse);
-                long personNow = num[0] * num[2] + num[3];
-                long forPersonNow = 0;
-                long totalPlayCount = 1;
+                double personNow = num[0] * num[2] + num[3];
+                double forPersonNow = 0;
+                double totalPlayCount = 1;
                 totalPlayCount *= num[1] * (num[1] - 1) / 2;
                 totalPlayCount *= num[0];
 
                 double powNum = Math.Pow(2, 65);
 
-                for (long i = 1; i <= powNum; i *= 2)
+                for (double i = 1; i <= powNum; i *= 2)
                 {
                     if (i >= personNow)
                     {
@@ -29,12 +29,10 @@ namespace hello
                         break;
                     }
                 }
-
-                for (long i = forPersonNow / 2; i >= 1; i /= 2)
+                for (double i = forPersonNow / 2; i >= 1; i /= 2)
                 {
                     totalPlayCount += i;
                 }
-
                 forPersonNow = forPersonNow - personNow;
 
                 Console.WriteLine(num[0] + "*" + num[2] + "/" + num[1] + "+" + num[3] + "=" + totalPlayCount + "+" + forPersonNow);
